myhaki/requirements.txt
asgiref==3.9.1
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
numpy==2.3.2
pgvector==0.4.1
pillow==11.3.0
psycopg2-binary==2.9.10
pyjwt==2.10.1
sqlparse==0.5.3
